refactor(shared): one signal-watcher primitive for CLI and desktop
The desktop renderer's shell watcher was a hand copy of the CLI's
`subscribeToSignalChanges` — same coalescing microtask, same re-arm in
`finally`, and a comment at the copy naming the CLI original as the source
of the defect it fixes. Two copies of the "a Watcher fires once" rule means
the next fix lands in one of them.

`subscribeToSignalChanges` moves to `@shared/signals` (which already exists
and which the desktop renderer already imports `Signal` from) and absorbs
the one thing the desktop copy did extra: reading each pending computed so
the watcher re-tracks its dependencies. That is a no-op for a watched
`Signal.State`, which never reports pending, so the four CLI call sites are
unchanged in behaviour.

The desktop's `shellDeps.get()` prime stays at its call site: an unevaluated
computed has no dependency set, so folding the prime into the helper — or
dropping it — would silently stop the shell re-rendering.
